Sync md and docx

+++ listings/ch11-writing-automated-tests/no-listing-04-bug-in-add-two/src/lib.rs
@@ -1,5 +1,5 @@
 // ANCHOR: here
-pub fn add_two(a: usize) -> usize {
+pub fn add_two(a: u64) -> u64 {
     a + 3
 }
 // ANCHOR_END: here
